pyproject.toml
[project]
name = "term-trace"
version = "0.1.0"
description = "Trace terminal sessions and generate structured summaries."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Jaafar Chakrani"}]
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "textual>=0.48.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
term-trace = "term_trace.cli:main"
write-jsonl-entry = "term_trace.scripts.write_jsonl_entry:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
